+++ backend/src/models/mod.rs
@@ -1,0 +1,11 @@
+pub mod company;
+pub mod company_card;
+pub mod feedback;
+pub mod file;
+pub mod image;
+pub mod layout;
+pub mod map;
+pub mod prepage;
+pub mod shortcut;
+pub mod tag;
+pub mod user;
