+++ backend/src/routes/shortcut.rs
@@ -1,41 +1,10 @@
 use actix_web::web::Json;
 use actix_web::{delete, get, post, put, web, HttpResponse, Responder, Result};
-use serde::{Deserialize, Serialize};
 use sqlx::PgPool;
-use strum_macros::{Display, EnumIter, EnumString};
 
+use crate::models::shortcut::ShortcutWeb;
 use crate::services;
 use crate::services::auth::AuthedUser;
-
-#[derive(Debug, Clone, Deserialize, Serialize, Eq, PartialEq)]
-pub struct ShortcutWeb {
-    pub id: Option<i32>,
-    pub name: Option<String>,
-    pub description: Option<String>,
-    pub link: Option<String>,
-    pub icon: Option<String>,
-}
-
-impl Default for ShortcutWeb {
-    fn default() -> Self {
-        Self {
-            id: Default::default(),
-            name: Default::default(),
-            description: Default::default(),
-            link: Default::default(),
-            icon: Default::default(),
-        }
-    }
-}
-
-#[derive(EnumIter, EnumString, Display, Debug, PartialEq, Eq, Hash)]
-pub enum RequiredField {
-    Id,
-    Name,
-    Description,
-    Link,
-    Icon,
-}
 
 pub fn routes(cfg: &mut web::ServiceConfig) {
     cfg.service(
