+++ backend/src/services/batch/tag_processor.rs
@@ -8,7 +8,7 @@ use super::{
     BatchProcessError, XlsxSheetProcessor,
 };
 use crate::{
-    routes::tag::{RequiredField, TagWeb},
+    models::tag::{RequiredField, TagWeb},
     services::tag,
 };
 use sqlx::{Pool, Postgres};
